Rename SignInGate/signInGate to SignInGate/signInGateMappings
