+++ dotcom-rendering/src/components/SignInGate/signInGateMappings.ts
